Way More Modular (WMM)

+++ include/settings.h
@@ -6,7 +6,7 @@
 /*   By: angagnie <angagnie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/03/06 11:07:04 by angagnie          #+#    #+#             */
-/*   Updated: 2016/03/06 11:11:01 by angagnie         ###   ########.fr       */
+/*   Updated: 2016/03/12 11:01:38 by angagnie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 ** double or float
 */
 
-typedef double  t_real;
+typedef float	t_real;
 
 # define DEFAULT_FOV 60
 # define DEFAULT_RATIO 16, 9
